requirements.txt
accelerate
codetiming
colorama
datasets
dill
flash-attn
flask
hydra-core
numpy
pandas
pybind11
ray
tensordict==0.7.2
transformers==4.51.3
vllm==0.9.2
wandb
torchdata==0.9.0
peft==0.15.2
tenacity==9.1.2
regex
# for accelerate inference speed
# pip install flashinfer-python -i https://flashinfer.ai/whl/cu124/torch2.6/
